+++ app/src/main/java/com/amalitech/bookmeetingroom/MainActivity.kt
@@ -9,7 +9,7 @@ import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import com.amalitech.core_ui.state.rememberBookMeetingRoomAppState
 import com.amalitech.core_ui.theme.BookMeetingRoomTheme
-import com.amalitech.core_ui.ui.BookMeetingRoomDrawer
+import com.amalitech.core_ui.components.BookMeetingRoomDrawer
 import com.amalitech.onboarding.login.LoginScreen
 
 class MainActivity : ComponentActivity() {
@@ -17,19 +17,20 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             BookMeetingRoomTheme {
-                val appState = rememberBookMeetingRoomAppState()
-                BookMeetingRoomDrawer(appState)
+                // TODO: uncomment this while testing, and move to appropriate position ...
+                // val appState = rememberBookMeetingRoomAppState()
+                // BookMeetingRoomDrawer(appState)
                 // A surface container using the 'background' color from the theme
-//                Surface(
-//                    modifier = Modifier.fillMaxSize(),
-//                    color = MaterialTheme.colorScheme.background
-//                ) {
-//                    LoginScreen(
-//                        onNavigateToHome = { /*TODO*/ },
-//                        onNavigateToForgotPassword = { /*TODO*/ },
-//                        onNavigateToSignUp = { /*TODO*/ },
-//                        onNavigateUp = { /*TODO*/ })
-//                }
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    LoginScreen(
+                        onNavigateToHome = { /*TODO*/ },
+                        onNavigateToForgotPassword = { /*TODO*/ },
+                        onNavigateToSignUp = { /*TODO*/ },
+                        onNavigateUp = { /*TODO*/ })
+                }
             }
         }
     }
