+++ core_ui/src/main/java/com/amalitech/core_ui/ui/BookMeetingRoomApp.kt
@@ -2,6 +2,7 @@ package com.amalitech.core_ui.ui
 
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import com.amalitech.core_ui.components.BookMeetingRoomTopAppBar
 import com.amalitech.core_ui.navigation.BookMeetingRoomNavHost
 import com.amalitech.core_ui.navigation.homeNavigationRoute
 import com.amalitech.core_ui.state.BookMeetingRoomAppState
